% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitize_iri.R
\name{sanitize_iri}
\alias{sanitize_iri}
\title{Sanitize a Text String for IRIs}
\usage{
sanitize_iri(x)
}
\arguments{
\item{x}{A string to convert into an IRI}
}
\value{
A string with removed illegal characters, expressions.
}
\description{
Sanitize a Text String for IRIs
}
